package.json:versions, html: order, index.js: default values and conditions
updates to the webpack and webpack-cli version.
changes to the structure of html, the game_state was put first then ai_status followed by the buttons MinMax and ANN, and last is the Undo and Reset buttons.
The script was linked to index.js
In index.js:
The draw_message and display_ai functions were changed
The default values of ai_type, white_ai, and black_ai were changed
The conditions in refresh_state were changed